LLM-Assignment-2/requirements.txt
pandas
scikit-learn
sklearn-crfsuite
joblib
nltk
textblob
matplotlib
